Refactor addressInput function to lib dir, write tests
